LibWeb: Split CalculatedStyleValue out of StyleValue.{h,cpp}
